[FIX] account_sepa: fallback if res is empty
* Go to Accounting > Sales > Payments
* Create a new payment
* Traceback! KeyError 'domain'

This happens because on creation, when calling super's
onchange_partner_id will return an empty dict, and at the end of the
method we assume that it has at least the 'domain' key in it, which is
not the case of course.

The solution is self-explanatory.

Introduced by 2c3b180

OPW-1830313