[FIX] website_version: see versionning changes
Website Versioning module (website_version) is applied when displaying
views by interfering when doing the "render the view of this `key`" and
showing the view with the same `key` but of the currently displayed
version.

The website.page model when directly rendering a page would directly
specify its `id`, thus website version was bypassed and just the master
page was displayed.

With this commit website.page ask to render the view by `key` if we are
on a given version.

This PR has the counterpart in community of odoo/odoo#23846

opw-1824957
closes #2018