[FIX] website_contract: no next next success mail
If the recurring_next_date is set, it should already be the one of the
next invoice (since the transaction is in done or authorized and the
period is incremented on the call back of _do_payment).

So we were getting the date of the invoice after the next.

opw-1824142
closes #2034