[FIX] web_studio: reset operations after Studio view edition
After editing the Studio view arch with the XML editor, the operations stack
needs to be reset to avoid conflicts.

This was broken since rev. odoo/enterprise@d13ddd2 because the `studio_view_id`
is reset on each change. However, one call (/web_studio/edit_view_arch) out of
the two possible ones does not return the view id so it was wrongly reset to
`undefined`.