[FIX] web_studio: correctly render graph editors
Backport from b53a3e210a79edfb36faf21f94aea91d32348838

This commit is a counterpart of a community commit making a change
in the way graph views are rendered. The problem is that the lib
used (nvd3) requires that the rendering of the graph is done into
the DOM (so that it can compute positions correctly). This is why
the graph rendering was done in a setTimeout. However, this recent
community commit removed the setTimeout, and used the
'on_attach_callback' hook instead (called when the widget is
attached to the DOM). The graph rendering is now done in that
function.

However, this function wasn't called in Studio, before this rev.