[FIX] account_reports: include adjustment taxes in tax report
odoo/odoo@57ecab7154fb2f63314650e7ce76925864382536 introduced a new
type 'adjustment'. When this type was used on a tax it would lead to a
traceback because there was no 'adjustment' key in `groups`.

This resolves that and also adds a more flexible way to get the string
of the value of the type_tax_use field instead of hardcoding it.

Backport of d282ce422cf72a608e5bef7e8f7522140c211b44 and partial
backport of related fix in 170ecabde00ec57d9cac834f7dba8285bfec1a05,
initially done in master.

Related to PR #2050