[FIX] account_reports: include adjustment taxes in tax report
odoo/odoo@57ecab7154fb2f63314650e7ce76925864382536 introduced a new
type 'adjustment'. When this type was used on a tax it would lead to a
traceback because there was no 'adjustment' key in `groups`.

This resolves that and also adds a more flexible way to get the string
of the value of the type_tax_use field instead of hardcoding it.

Part of task 38862
Part of the PR #2050