[FIX] marketing_automation: Process participants/traces by batch
For big marketing campaigns it's impossible to manage all the participant/traces creation/execution
in a single transaction. This commit splits the data creation/execution into several batches that are
commited during the cron call.

That way, if a transaction can't finish due to a timeout, the created data is consistent and won't
be processed twice, allowing the processing of other data.