[FIX] sale_subscription: display meaningful name in portal
As of 11.0 subscriptions do not always have an analytic account parent
and the corresponding derived name. The `name` field does not contain
very useful information, usually only the name of the subscription
template.
On the portal, showing the `display_name` is more useful, as it includes
the contract code and the name of the customer.

This small change required a fix in the name_get() method of
subscriptions, as user do not always have read access to the res.partner
assigner to the subscription. This extra sudo() is considered safe as it
only returns the name of the related partner - similarly to the normal
behavior for linked m2o records.