[IMP] web_dashboard: additional measures on subviews
Related to commit in community that removes many2one fields in graph and pivot
views. Reason of this commit is to not have a huge list of measures.

With this commit, all many2one fields in the arch of a dashboard are correctly
passed to subviews for measures. That way, only some relevant many2one fields
are considered as measures in subviews.