[FIX] web_dashboard: propagate m2o measures in sub views
Before this fix, the code was incorrectly taking the name attributes for
the m2o aggregates as additional measures for the graph/pivot views.

This was wrong because the aggregate can have its own name, independant
of the field name.  As a result, the m2o measures were simply not
displayed in the sub views (even though the rest of the ui still works).