[FIX] stock_barcode: lot wizard errors management
- Multiple scan on a same serial number was not handle
- Quantity done summary was not updated

It happens because the check on multiple serial number is called
inside on_barcode_scanned. However on_barcode_scanned is only call
when a candidate is not found (thus is not duplicated).

qty_done was not handle at all, in order to correct it we use an onchange.
lot_barcode_handler thus should notify the server in order to trigger
onchange that could result in a small performance loss. It could
be improved on master by including the onchange behavior inside
the lot widget and only interact with the server on save.