[IMP] timesheet_grid: encode in any 'time' UoM
This commit extend the new defined mecanism in
community edition to allow users to encode their
timesheet in days. For each time UoM, a different
js widget is added on unit_amount field to ease
the encoding. Views defined in database are prior
over the field_view_get customizations.

Also, this commit add the UoM name as additionnal
string on Total row/column of the grid view.

Task #39079