[ADD] quality_iot, quality_mrp_iot: links between quality points and iot machines + measure widget
In the module quality_iot, we add a link between quality points and iot devices.

In the module quality_mrp_iot, we add related fields on the workorder
linking to the device of the current quality check's control point.
For the 'take a measure' quality step, we change the widget of the input field
in order to be able to get the measure from the iot box.