[FIX] l10n_mx_edi: Calculate correctly taxes when included in the price
Currently, when the CFDI is generated, all taxes are included assuming
the amount is not included in the product price, causing the base price
and tax amounts to be calculated incorrectly when the tax is actually
included in the price.

This change causes all prices to be calculated using the Total and
Subtotal amounts, instead of the Unit Price, to avoid misscalculations
in cases like this.