[IMP] account_reports: filter on partner (category) in financial reports
This commit allows to define an class variable filter_partner on financial reports that could be used to filter the report by partner or partner category (in the same fashion as for the analytic accounts/tags).

This option was also enabled for all the partner reports (aged receivable/payable, partner ledger).

Was PR #1928
Was task: 36626