[FIX] web_enterprise: Do not block screen readers to navigate over home page (#1990)
Currently, it's not possible to navigate over the main menu using a screen reader (for visually impaired), because all key presses are redirected to the search text box,
even when it's not being used. Those key presses include the keys
<Tab> and <Shift>, which are essential when navigating using the
keyboard.

This change causes that, when the search text box is not being used
(i.e. when it's empty) the keys <Tab> and <Enter> are not captured
and invalidated.

In addition, this also prevents the keys <Shift>, <Ctrl> and <Alt>
from being invalidated, because, besides being needed, it makes no sense
to redirect them to the search text box, only character keys are useful.

The only functional workaround for blind people consists to create a bookmark dir in the browser to save all direct URL, and this is not a good way to manage it.