[IMP] web_studio: add `move` positions in normalization
Before this rev. the view normalization was only creating `replace` or `add`
xpaths while parsing the diff between the archs with and without the Studio view.

This approach was not optimal for moved fields, mainly because all attributes
were copied on the `add` xpath. If the Studio view was created before installing
a new module that modified a attribute for example, this one would have been
lost in the process.

With `move` xpath, the entire node (and its attributes) are moved.

The normalization thus needed to be able to spot these moved fields and create
the corresponding xpaths. To do so, we spot fields appearing in both '-' and '+'
lines in the diff.

This heuristic however is not complete. Imagine the following situation where a
field in a group is moved in the upper group. The diff, instead of writing '-'
and '+' for the field, will move the boundary of the group.

Like
...
  /data/form/sheet/group/group/field[@name=website]
  /data/form/sheet/group/group/field[@name=website][@closed]
- /data/form/sheet/group/group[@closed]
- /data/form/sheet/group/group
  /data/form/sheet/group/group/field[@name=function]
  /data/form/sheet/group/group/field[@name=function][@closed]
+ /data/form/sheet/group/group[@closed]
+ /data/form/sheet/group/group
...

The normalization thus needs to spot these cases where the group boundaries are
moved and declare all the fields inside as moved fields.

Moving boundary algorithm is a joint work with @dmo-odoo.