[IMP] web_studio: allow custom monetary field creation
Before this rev. it was not possible to create monetary field on a custom model.
A monetary field requires a currency field on the model (named `currency_id` or
`x_currency_id`) to be created. An error was thus displayed on custom models (or
existing models without a `currency_id` field).

A dialog is now displayed at the monetary field creation to propose to create a
currency field first.

Note that the currency field must be present in the view for the monetary field
to be correctly displayed.