[FIX] l10n_mx_edi: Adjust permissions in tests
Starting from:
ba25420312b1
The attribute `uid` in test clases is now a property, and it's binded to
`env.user`.

Because of that, the behavior when using `self.uid` has changed, so
changes need to be performed.

This commit performs all needed changes in order to comply with this new
vehavior.