[FIX] helpdesk: close activities when a ticket is done.
Starting in saas-11.3 automated activities are created for tickets based
on SLA deadline. However those are not closed when ticket reaches a close
stage. This commit fixes that. Putting a ticket in a closing stage now
set automated activities as done.

This commit is done after feedback from internal testing phase of saas-11.3 .