[FIX] l10n_mx_edi: Fix cache access inconsistencies
Since invoices are created in the tests using a different user, they get
its own new environment. Moreover, since they are under a test class,
the cache is not automatically invalidated.

Because of the above, when an invoice is created and a refund is created
for that invoice, the invoice's values and the refund's values related
to that invoice may differ, leading to inconsistencies depending on
which one the information is retrieved from.

This change ensures the cache for the refund to be refreshed before any
operation is done, to ensure both the invoice and the refund refers to
the same information.