[FIX] web_studio: fall back to non-normalized view
Previous to this commit, if an error occurred during the normalization,
a ValidationError would be raised and shown to the user.

This is not expected behavior, the normalization is a "nice-to-have" but
it should not, in any case, deter the functionality of studio.

A fall back was in place with a try/except clause, however the except
was expecting a ValueError when in reality it only gets
ValidationErrors.