[REF] mail_push: refactor mobile notification code and move it to message model
Mobile notification code is located on channel model although it it totally
not related to channels. It is actually linked to message notification
process.

This commit update code so that it is now an override of _notify method
of mail.message. This will allow to clean a bit the code to improve
performances and will ease future improvements in notification process.

In this commit some code is cleaned to lessen impact in term of performance
of mail_push module in mail-related performance tests. Indeed mail_push is
automatically installed and impacts tests located in test_mail module. As
shown in upcoming community commit this enterprise commit has an impact on
top counters and allow to lessen impact of enterprise modules on mail
performance.

This commit is linked to task ID 1838693. Closes #2111.