[IMP] web_studio: use `extra_trigger` in tour
To verify that a new added field was correctly added in the view, extra
steps (clicking on the field) were performed in the tour.

The `extra_trigger` was designed for these precise use cases and the extra
steps uselessly extend the tour duration (rpcs are done to fetch the default
value when clicking on a field).