[FIX] l10n_mx_edi: Support write_off in payment
When the payment CFDI is generated, must be indicated the amount paid
(Monto), and must be itemized which invoices are paid with the payment.
But the sum in the invoice detail not must be graten that the amount
total.

When a payment have write_off, the invoice is marked as full paid, but
in the CFDI generation must be indicated only the amount paid.

Without this fix, in the next case I have the next error:
La suma de los valores registrados en el campo ImpPagado de los
apartados DoctoRelacionado no es menor o igual que el valor del campo
Monto.

Invoice total: 1,026.60
Amount Paid: 1,026.00
Write Off: 0.60

This was fixed considering the write off in the payment.

Also was improved that not round the amount when get the rate to the
invoice detail, to avoid bad calculation.

Closes #2099