[FIX] currency_rate_live: log instead of triggering UserError when the cron updating rates cannot connect to a provider
We still trigger errors when the user tries to manually update the rates for a deprecated provider (Yahoo) or one that is temporarily unavailable. For that, I had to add a return value to update_currency_rates (I got qdp's approval for that).

Before this fix, if one of the companies had a deprecated rate provider set, or could not connect to its provider for some reason, the cron directly raised an error, and obviously stop irating over the companies. It caused the ones that were not yet treated not to receive any update to their currency rates, even if they were using a different provider.