[FIX] account_reports: wrapping of column headers
A set a commits modified the behavior of header's wrapping in reports:
23ec9e01e37f66d1a870c2df34847550473684e7
a2347c095b0bf8889ca99f1b1a79c5ed2cff713a
9cf454229ab31291d9f54a906aa8b816b19f5285

The last commit forces a `white-space: nowrap` value, but this
reintroduces the issue originally solved with the very first commit.

We set the value back to `normal` (since it was the behavior until not
so long ago), and instead manually set a `nowrap` value when necessary.

opw-1834431