[FIX] web_gantt: duration_unit from 'hours' to 'minutes'
For the moment, durations in web_gantt are calculated in hours,
and values are rounded. For instance, the duration between 08:00
and 09:26 is 2 hours.

This setting introduces some strange behaviours: a parent task[1]
has a duration in hours that is rounded, while children tasks have
a duration in seconds.

Example:
    parent-task: Mo/00001         08:00 - 10:00 (2 hours)
     child-task: Manual Assembly  08:00 - 09:26 (1 hour 26 minutes)

This should not be the case, but at least the child-task has the correct
start and end datetime values.

When draging any of the children tasks somewhere else in the gantt,
its duration becomes computed in hours.

Example: when dragging "Manual Assembly" so that start date is 10:00
    parent-task: Mo/00001         10:00 - 12:00 (2 hours)
     child-task: Manual Assembly  10:00 - 12:00 (2 hours)

Which seems incorrect, has we should expect the child-task to keep its
duration with a simple time translation.

This commit suggests a not-so-good solution, which consists of making
any computation in gantt in 'minutes', instead of in 'hours'.

Known issues:
- The datetime of a child-task contains seconds, so dragging this task
  may still result in a lost of around 1 minute.
- The Gantt subscales of day view has an accuracy of hours. When moving
  an element, it will start at the beginning of an hour. Moreover,
  resizing an element can only be done in multiple of hours. Setting the
  unit to 'minute' makes the view unreadable, even with a larger step.

opw-1839520