[IMP] l10n_mx_edi: Adding fields on report invoice adapted to CDFI3.3
On the CFDI 3.3 mexican standard were added new attributes and elements
that have to be placed on the PDF invoice in order to make the printable
version of the invoice show that content present on the XML signed file.

The new report invoice format shows on the PDF the Usage, the product's
SAT Code, the SAT Unit and the Emitter RFC when required.