[FIX] website_contract: use determinist company for recurring renewal
The company used to test the subscription automatic payment
renewal was determined using a search with an empty domain
and without order.

According to the modules installed and their order,
the company used could be different
(especially after installing the multi company module).

This revision makes sure to always use the same company,
the one of the user.

This is important to test the taxes,
as the taxes computation filters out taxes which are
not in the same company then the invoice.