[FIX] website_calendar: tests, administrator is on sick leave
When the module `hr_holidays` was installed before `website_contract`,
the test `test_extreme_timezone_delta` failed because `hr_holidays`
adds 3 sick days to the administrator as a demo data,
from 20 to 22 of the current month,
and if the 20, 21 or 22 of the current month was a Monday,
the test failed because it checked that all Monday "working slots"
were well assigned to the administrator,
which coudln't happen as the administrator was sick. What a shame.

The issue is solved by creating user dedicated to the test,
to make them demo data independent.