[FIX] web_studio: attribute removal remove it
When an attribute is removed in studio, the normalization is eg.:

  ```
  <xpath expr="..." position="attributes"/>
  ```

instead of an expected:

  ```
  <xpath expr="..." position="attributes">
    <attribute name="attribute_name"/>
  </xpath>
  ```

This is alright if the attribute was just changed, but if it was removed
the removal would not be applied.

With this change, the following node:

 `<b string="world" class="h"/>`

would be stringified to:

/b
/b/attributes[@name=class]
/b/attributes[@name=class]/attribute[@name=class][@text=h]
/b/attributes[@name=class]/attribute[@name=class][@text=h][@closed]
/b/attributes[@name=class][@closed]
/b/attributes[@name=string]
/b/attributes[@name=string]/attribute[@name=string][@text=world]
/b/attributes[@name=string]/attribute[@name=string][@text=world][@closed]
/b/attributes[@name=string][@closed]
/b[@closed]

So we have an wrapper `<attributes name="attribute_name" />` tag for
each node that allow to differentialte between attribute removal and
modification.

opw-1838786
closes #2168