[FIX] timesheet_grid: reminders nextdate computation
For weekly/monthly reminders the next week/month was wrongly computed,
it sets the nextdate to one week/month before instead.

The whole mechanism is quite dodgy, datetime.now() could very well not
at be the same date as the start of the cron's execution and we could
do without the nextdate field by having the cron check if today is a
valid timesheet reminder execution day. But this refactoring has to be
for future versions.