[IMP] helpdesk: implement digest KPIs / tips
Purpose of this commit is to add closed tickets KPIs in digest emails.
Those are added in the default digest template.

A tip to remind to use the mail gateway is also added.

See commit https://github.com/odoo/odoo/commit/8eec939a7885c860c6c821298709e567bd625027 in community for more information.

This commit is linked to task ID 30655 and PR #2180.