[FIX] l10n_mx_edi_cfdi_33: Solved some cases with multi currency
- The 'Payment rate' attribute could have 6 decimals, and the amount
not must be rounded.

The compute method in res.currency by default has the attribute round as
True, but We do not need the value rounded to has a calculate more
precise.

Example:
* Without this change in the XML
  TipoCambioP="19.090000" Monto="157.47"
  ImpPagado="3006.72"

* The calculate:
  19.090000 * 157.47 = 3,006.10

The PAC validate that the amount calculated is higher than ImpPagado and
3,006.10 is not higher that 3006.72.

* With this change:
  TipoCambioP="19.094172" Monto="157.47"
  ImpPagado="2958.52"

* The calculate:
  19.094172 * 157.47 = 3,006.75926

With this the CFDI is signed.

========================================================================

- The amounts in ImpSaldoAnt, ImpPagado and ImpSaldoInsoluto must be
expressed in the same currency, to that is is a currency different to
MXN, must be used the amount_currency.

========================================================================

Adapted the expected CFDI to the new change that support 6 decimals, and
call the method compute_taxes() before that the invoice is validated to
avoid error in the test.