[FIX] web_grid: issue with focus/reloading with wrong context
Before this commit, editing a value in the grid view was a frustrating
experience.  For example:

- open timesheet grid view
- go back to previous week
- add one line, if there are none
- focus one cell
- change a value, press tab

Then, the value is changed, but the grid view reloads the current week
(so, not the week you were previously), and put the focus on the search
view.

The bug was introduced in saas-11.2, with the action manager
refactoring. This is caused by various small issues:
- to edit a cell, the grid view used a execute_action rpc, which is now handled
 slightly differently, and returns an action with the session context
- the result of that execute_action is fed into the reload handler
- the reload handler replace the grid current context with the session
  context, which does not contain the correct grid_anchor
- the grid view then reload the view, which now causes a
  updatecontrolpanel call, and this is why the focus is now on the
  search view.

To solve these issues, we now directly call the adjust_grid method on
the correct model instead of using the execute_action indirection. This
solves the context problem, but not the focus issue.  We also do not
reload the view, but simply reload the data and upate the grid. This
allow the grid view to not call updatecontrolpanel, and is also slightly
more efficient.