[ADD] l10n_mx_edi_payment: Save the expedition date to payment task#24972
The CFDI payment complement have an attribute to the expedition date,
that indicate the date when is generated the CFDI. That value can not be less
that 48 hours, compared with the current time.

Initially that value is taken from the payment date, but this is not
right, because the payment date is when I receive the money
in the bank, but I could emit the CFDI in other date.

The SAT indicate that the payment complement could be emitted 10 days
after of the month when was received.

http://www.sat.gob.mx/informacion_fiscal/factura_electronica/Documents/Complementoscfdi/PregFrec_RP.pdf

[REF] l10n_mx_edi_payment: mission utf8, translation, avoiding for on method, refactoring manifest