[FIX] timesheet_grid: reminder template rendering dates
Before this commit the dates were wrongly computed. For weekly
intervals it took the start of the week of when the cron was
executed instead of the start of the previous period. For both
weekly and monthly periods it always ended at today, instead of
the end of the previous period.