[IMP] web_grid: display measure label
The grid view displayed number in cells, but this is difficult to
know the meaning of those numbers. Like in timesheet or forecast, only
the widget can help knowing if the number are expressed in Hour or in
Days.

This commit provide a mecanism to display a label beside the Total
word on final row and column. This will help users to know what they are
manipulating.