[FIX] web_studio: avoid pushing state
In Studio, the 'intern' actions (i.e. Reports, Filter Rules, etc.) should not
push their state because we want to keep an URL that can be refreshed. However
this was not the case before this rev. (for example, by updating the search
view in 'Reports' Studio action).

In a classic workflow, the flag `pushState` is always reset after a `do_action`
in the action manager, which is a behaviour we want to prevent in Studio, hence
this override.

The use of `done` in web_enterprise is there to avoid loosing `action` while
chaining the deferred.