[FIX] barcodes_mobile: avoid virtual keyboard while scanning barcode
In mobile, the virtual keyboard is automatically openend when an
input element has the focus.
As we don't need it, the keyboard was closed right after opening it.
This behavior was a bit weird for users.

The virtual keyboard is now directly closed when it appears in the
mobile app. So, the user wouldn't see it.

Note that the barcode input is not removed from the DOM anymore,
so the focus on the input is kept at each scan. See odoo/odoo@a30d9aa

We had to create a new module for this fix because web_mobile.rpc
is available in enterprise only.