[FIX] timesheet_grid: removal of required constraint
The description is required on an analytic line. timesheet_grid modul
tries to remove this constraint, as when ajusting timesheet from the
grid view, this is not possible to set a description.
Removing this required constraints brings error in log file, when
updating analytic (which is trying to restore the required
constraint on potential NULL description line).
This commit provides a default description when adjusting timesheet
from the grid view, in order to remove the removal of the
constraint.

Closes #22097