[FIX] timesheet_grid: employee reminders are all the same
`hr.employee`'s method `get_timesheet_and_working_hours` generates a
template result with THE SAME dictionary for every employee_id. So when
we fill the dict, every iteration changes the value in the shared dict
updating it for all employees.

The fix is obvious: generate a new dict for each employee_id.

opw: 1850369