[FIX] account_reports: do not display full_reconcile liquidity lines
Make one Payment
Reverse the entry of this payment.

Before this commit, only the lines in the receivable were reconciled.
The lines in the liquidity were'nt, leaving the two entries appearing
in the reconciliation report

After this commit, the four lines are reconciled two by two, even in the liquidity
account.
The report do not display those lines anymore.

OPW 1816641

ref odoo/odoo#25026
closes #2234