[FIX] web_studio: menu moved is saved
Currently in web studio when editing menus, moving a menu somewhere else
would only happen if the index of the menu was changed.

This is because the library (https://github.com/ilikenwf/nestedSortable)
when deciding if a menu has been moved check:

- the index of the menu item,
- the ID attribute of the original parent and current position parent

but odoo did not specify IDs thus only the index could differ.

There was also the issue that doing several menu moves or deletions are
only saved on confirm, but editing or creating a new menu reload the
modal losing the temporary changes.

With this commit the temporary changes are saved when a menu has been
added or edited.

Without the fix, the added test failed with:

  "Cannot read property 'sequence' of undefined"

note: 11.0 equivalent of 10.0 #2244

opw-1850527
closes #2242