[FIX] web_studio: menu moved if index conserved
Currently in web studio when editing menus, moving a menu somewhere else
would only happen if the index of the menu was changed.

This is because the library (https://github.com/ilikenwf/nestedSortable)
when deciding if a menu has been moved check:

- the index of the menu item,
- the ID attribute of the original parent and current position parent

but odoo did not specify IDs thus only the index could differ.

note: the same fix is done in 11.0 (#2242) with a test

opw-1850527
closes #2244