[FIX] web_dashboard: search view do not overwrite subviews contexts
By the past, if a user would change the graph subview measure
(or the pivot subview measures) and then deactivate a filter from
search view, the effect would be that the measure selection
would be reset at the last value known by the dashboard controller.
This commit fixes that situation.