[FIX] web_studio: make runbot green again
There is a long-standing bug (KeyError: ir.ui.menu.display_name, caused by a registry
issue with multiple workers) on runbot that is annoying most Odoo devs. It thus has
been decided that the tour should be temporarily modified to move the steps causing
the error at the beginning of this test to avoid the registry to be reloaded
before the write on ir.ui.menu.

This rev. should be reverted once the bug is fixed in the ORM.