[FIX] web_grid: display field values in groupby by a non-relational field
When grouping by a non-relational field in a grid view (e.g. by name),
the row title would only show one letter of the field.
This is because relational field are given in the form of a couple (id, value)
whereas non-relational fields are simply given as value.
The code simply extracted the value by accessing v[1], which sadly returned "a"
for "name".
We now check the type to always return the correct value.

opw 1854736