[IMP] l10n_mx_edi_cfdi_33: Some improves
Recently the PAC added the validation that when the VAT has a generic
value the in 'ResidenciaFiscal' must be empty. According to the SAT
documentation:

>> Si el RFC del receptor es de un RFC registrado en el SAT o un RFC
>> genérico nacional, no se debe registrar este atributo.

From:
http://www.sat.gob.mx/informacion_fiscal/factura_electronica/Documents/cfdv33.pdf
Pag. 67.

Now validate that the VAT is not generic when set the value in that
attribute.

------------------------------

Validate the write_off is found in the payment, because on some cases
generate traceback because write_off is assigned as float.