[FIX] helpdesk: auto-reply template
- Configure an email alias so that a new email creates a new helpdesk
  ticket.
- Set a mail template on the first stage of helpdesk
- Send an email to the alias => a ticket is created, and the mail
  template is received.
- Reply to template received.

The 'Reply-To' address is set to the 'catchall' address, while it should
be the alias. The email gets nowhere.

By setting `composition_mode=mass_mail`, the template sent goes through
`self.env['mail.thread']...message_get_reply_to` in `get_mail_values`,
which bypasses `message_get_reply_to` overridden in `helpdesk.ticket`.
Moreover, in this mode, a copy of the email is not posted on the
original ticket. If the template is set as `auto_delete`, a reply won't
be able to find its way.

There is no real need for `composition_mode=mass_mail` in this case.

Backport of ff172174c35b58e1a939a08447cf548f74590392

Note: in v10, this also solves the issue that the `email_from` is empty
=> the mail template cannot be sent.

opw-1849786