[IMP] website_sign: make email templates bioutifoul
Purpose of this commit is to enhance quality of templates proposed by Odoo
and make them use notification layout when send by emails. Those emails
are cleaner and more up to date compared to other emails.

Including
  * complete rework of the strange template used in website_sign. It is
    now split into 3 templates as it has basically three different
    uses. It is also moved to qweb as they are not really email templates
    but more templates used to send content to people. They are not meant
    to be used outside of a method preparing data for it;
  * write a specific method for the double rendering (one for the content
    template, one for the layout template);

Please note that website_sign is a bioutifoul application and I loved
working in it.

This commit is related to task ID 51122 (and PR #2060).