[IMP] *: review date type groupbys in search views archs
The new possibility to select a specific interval for any
date or datetime field in the groupby menu has been added.
This makes some filter akward, for instance "Order Month".

Accordingly, we have rewritten the string and help attributes in
search views archs and removed groupbys related to the same field
(that would have become redundant).