[FIX] mrp_account: tests
First, as the produc lines are filled by an onchange, call it.
Secondly, before rev [1], if a product to consume was tracked, the
default value was 0 for qty_to_consume and the test didn't filled it.
Now, the onchange set the qty_to_do with the same value as
qty_to_consume, so a serial was required. We chose to disable the
tracking for this tests at the moment.

Now that the previously skipped products are used (because they aren't
tracked anymore), the first assert on the move value should be adapted.

We'll fix the second assert in a following commit.

[1] 8c524b265a3632a038d1cae451291cb263f2c5dd