[FIX] account_reports: bank reconciliation report: use correct signs to list unreconciled payments
This is actually a fix of an error introduced by commit https://github.com/odoo/enterprise/commit/39e24cfaae2f6928409e4852d709850d68d1f3cd . This commit changed the domain when searching for payment move lines, so that we only consider the lines made on accounts of type 'bank'. But doing so, we ended out selecting the lines balancing the lines we did select before, causing a sign inversion on the balance.