[IMP] web_enterprise: allow borders on list_view CSS
Before this change it was not possible to add borders on sale order lines.

It is now possible.

The removed border from the tfoot should not be an issue, because it already wasn't displayed due to the collapse setting.

The border on the tfoot actually displayed is done with a shadow.