[FIX] account_accountant: correctly set bank and cash settings
Currently settings for account accountant bank and cash are added by
classname 'app_settings_block'. But there are so many classes with this name
that by default it takes first class. First found application receives this
settings.

This commit fixes that by correctly putting the settings in the invoicing
settings.

Related to task: 1858217
Closes PR #2296.