[IMP] account_sepa*: use account.batch.payment to collect/manage SEPA payments
Business flows for SEPA credit transfers and direct debits have been unified and now use the batch payment objects. This allows a better end-user experience and also to use the former 'batch deposit' feature to collect all the payments made to your suppliers at once, if the bank made a single entry for that in the bank statement.

Was task 35558
Was PR #1844