[REF] stock_barcode: client action: tracked line
a tracked line should always display the barocde icon so the user knows
what he is supposed to do

we slightly refactore "setLotName". actually this method is called even
when scanning a lot_id. data are written on the right fields but for
ease we use the same logic for both cases. now this method always
replace the existing span.