[FIX] quality_mrp: Wrong UOM on workorder with lot
With a BOM, the uom of a BOM line can be different of the uom of the product
The qty_done is computed in function _create_checks on model 'mrp.workorder' with the
uom of its relative workorder's move_raw_ids. So the component_uom_id must be in this
uom too. As the move_line_id is computed in function _next on model 'mrp.workorder' with
its relative workorder's move_raw_ids, the uom of move_line_id can be used.

opw:1858002