[FIX] account_yodlee: display success message also on partial success
When a synchronisation returns the status PARTIAL_SUCCESS (one of the account
failed to sync), the status is PARTIAL_SUCCESS instead of SUCCESS.
The js code was trying to load the template YodleeLoginTemplate (used for MFA)
which failed as not all the data was available (providerAccount.loginForm is
undefined).

With this commit, a success message will be displayed for the accounts that
succeeded to sync.
In case the failing account is linked to a journal, another error would be
raised at the server side.
This commit does not hide errors.

Details about the partial success is still available on the form view of the
online sync record.

opw-1856875

closes odoo/enterprise#3660