[ADD] web_unsplash: Find free high-resolution images from Unsplash
This commit adds a new module `web_unsplash` that adds the possibility in the
media dialog to search high quality images on Unsplash.com.

They were not enough images in the media dialog (image widget).
We were obviously not covering all the business.

Unsplash API requests are limited to 50/hour with test accounts and 5.000
per hour with production account.

== Settings ==
- From `General Settings` you can enable the module
- Once enabled, you can add your unsplash key at the same place

== Front end (Media dialog) ==
- There will be a new input to enter a search query on Unsplash
- Unsplash results will then be displayed as normal attachment are
- Selecting image(s) will create an ir.attachment and the attachment
  URL will then be used in the image src

- In case you are trying to use the functionnality without having a key set,
  it will show a prompt to enter a key easily.
- In case you reached the request limit, it will tell you to wait or to
  upgrade your account.

task-35060