[FIX] mass_mailing_themes: bold/tech keep ok width
When editing mass mailing blocks, `box-sizing: border-box` is the
default thanks to CSS, thus when we specify on the logo+social header:

  have 20 pixels padding and have a 600 pixels width

we see a block of 600px with a content of 560px (between a left and
right padding of 20px).

But the box-sizing property is stripped when saving the mail, thus we
instead have a 640px block (600px content + 2 * 20px for padding) that
is expected to fit inside a 600px block: it doesn't.

It also seems that `border-box` is not fully supported, so this commit
set the CSS for it to work with `box-sizing: content-box` or if it is
stripped.

opw-1852839
closes #2317