[FIX] sale_subscription: handle subscription with multiple lines with the same product
It was assumed that a subscription could have only one line with a specific
(product_id, uom_id) couple.
In case where it is not the case, there is an ambiguity when upselling such
lines.
To avoid simpling adding to a random line, in case of ambiguity we add a new
line for the order.

To do so, we duplicate a random line while updating the description and quantity

opw 1859806